feature complete implementation of death action command
