+++ BloodNight-core/src/main/java/de/eldoria/bloodnight/command/bloodnight/ManageDeathActions.java
@@ -1,43 +1,16 @@
 package de.eldoria.bloodnight.command.bloodnight;
 
+import de.eldoria.bloodnight.command.bloodnight.managedeathactions.ManageMonsterDeathActions;
+import de.eldoria.bloodnight.command.bloodnight.managedeathactions.ManagePlayerDeathActions;
 import de.eldoria.bloodnight.config.Configuration;
-import de.eldoria.bloodnight.config.worldsettings.deathactions.MobDeathActions;
-import de.eldoria.bloodnight.config.worldsettings.deathactions.PlayerDeathActions;
-import de.eldoria.bloodnight.config.worldsettings.deathactions.PotionEffectSettings;
-import de.eldoria.bloodnight.config.worldsettings.deathactions.subsettings.LightningSettings;
-import de.eldoria.bloodnight.config.worldsettings.deathactions.subsettings.ShockwaveSettings;
-import de.eldoria.bloodnight.core.BloodNight;
 import de.eldoria.bloodnight.util.Permissions;
 import de.eldoria.eldoutilities.conversation.ConversationRequester;
-import de.eldoria.eldoutilities.core.EldoUtilities;
-import de.eldoria.eldoutilities.inventory.ActionConsumer;
-import de.eldoria.eldoutilities.inventory.ActionItem;
-import de.eldoria.eldoutilities.inventory.InventoryActions;
-import de.eldoria.eldoutilities.items.ItemStackBuilder;
-import de.eldoria.eldoutilities.messages.MessageChannel;
-import de.eldoria.eldoutilities.messages.MessageType;
 import de.eldoria.eldoutilities.simplecommands.EldoCommand;
-import de.eldoria.eldoutilities.simplecommands.TabCompleteUtil;
-import de.eldoria.eldoutilities.utils.ArgumentUtils;
-import de.eldoria.eldoutilities.utils.DataContainerUtil;
-import de.eldoria.eldoutilities.utils.Parser;
-import org.bukkit.*;
+import net.kyori.adventure.platform.bukkit.BukkitAudiences;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
-import org.bukkit.inventory.Inventory;
-import org.bukkit.inventory.meta.PotionMeta;
-import org.bukkit.persistence.PersistentDataContainer;
-import org.bukkit.persistence.PersistentDataType;
 import org.bukkit.plugin.Plugin;
-import org.bukkit.potion.PotionData;
-import org.bukkit.potion.PotionType;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
-import java.util.List;
-import java.util.Map;
-import java.util.OptionalInt;
 
 public class ManageDeathActions extends EldoCommand {
     private final Configuration configuration;
@@ -48,8 +21,10 @@ public class ManageDeathActions extends EldoCommand {
         this.configuration = configuration;
         this.conversationRequester = conversationRequester;
 
-        registerCommand("monster", new ManageMonsterDeathActions(plugin, configuration));
-        registerCommand("player", new ManagePlayerDeathActions(plugin, configuration, conversationRequester));
+        BukkitAudiences bukkitAudiences = BukkitAudiences.create(getPlugin());
+
+        registerCommand("monster", new ManageMonsterDeathActions(plugin, configuration, bukkitAudiences));
+        registerCommand("player", new ManagePlayerDeathActions(plugin, configuration, conversationRequester, bukkitAudiences));
     }
 
     // <monster|player> <world> <field> <value>
@@ -68,631 +43,5 @@ public class ManageDeathActions extends EldoCommand {
         }
 
         return super.onCommand(sender, command, label, args);
-    }
-
-    private static class ManageMonsterDeathActions extends EldoCommand {
-        private final Configuration configuration;
-
-        public ManageMonsterDeathActions(Plugin plugin, Configuration configuration) {
-            super(plugin);
-            this.configuration = configuration;
-        }
-
-        @Override
-        public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String label, @NotNull String[] args) {
-            Player player = (Player) sender;
-
-            World world = ArgumentUtils.getOrDefault(args, 1, ArgumentUtils::getWorld, player.getWorld());
-
-            if (world == null) {
-                messageSender().sendLocalizedError(sender, "error.invalidWorld");
-                return true;
-            }
-
-            String field = args[1];
-            String value = ArgumentUtils.getOrDefault(args, 2, "none");
-
-            MobDeathActions mobDeathActions = configuration.getWorldSettings(world).getDeathActionSettings().getMobDeathActions();
-
-            if ("lightning".equalsIgnoreCase(field)) {
-                LightningSettings lightningSettings = mobDeathActions.getLightningSettings();
-
-                Inventory inventory = Bukkit.createInventory(player, 18, "Manage Lightning Settings");
-
-                InventoryActions actions = EldoUtilities.getInventoryActions().wrap(player, inventory, e -> {
-                    configuration.save();
-                    sendMobDeathActions(player, mobDeathActions);
-                });
-
-                NamespacedKey valueKey = new NamespacedKey(getPlugin(), "valueKey");
-                actions.addAction(
-                        new ActionItem(
-                                ItemStackBuilder
-                                        .of(Material.LEVER)
-                                        .withDisplayName("Lightning active")
-                                        .withLore(String.valueOf(lightningSettings.isDoLightning()))
-                                        .withNBTData(
-                                                c -> c.set(
-                                                        valueKey,
-                                                        PersistentDataType.BYTE,
-                                                        DataContainerUtil.booleanToByte(lightningSettings.isDoLightning())))
-                                        .build(),
-                                3,
-                                ActionConsumer.booleanToggle(valueKey),
-                                stack -> {
-                                    boolean fieldValue = DataContainerUtil.byteToBoolean(DataContainerUtil.compute(stack, valueKey, PersistentDataType.BYTE, s -> s));
-                                    lightningSettings.setDoLightning(fieldValue);
-                                })
-                );
-
-                actions.addAction(
-                        new ActionItem(
-                                ItemStackBuilder
-                                        .of(Material.LEVER)
-                                        .withDisplayName("Thunder active")
-                                        .withLore(String.valueOf(lightningSettings.isDoThunder()))
-                                        .withNBTData(
-                                                c -> c.set(
-                                                        valueKey,
-                                                        PersistentDataType.BYTE,
-                                                        DataContainerUtil.booleanToByte(lightningSettings.isDoThunder())))
-                                        .build(),
-                                5,
-                                ActionConsumer.booleanToggle(valueKey),
-                                stack -> {
-                                    boolean fieldValue = DataContainerUtil.byteToBoolean(DataContainerUtil.compute(stack, valueKey, PersistentDataType.BYTE, s -> s));
-                                    lightningSettings.setDoThunder(fieldValue);
-                                })
-                );
-
-                actions.addAction(
-                        new ActionItem(
-                                ItemStackBuilder
-                                        .of(Material.BLAZE_ROD)
-                                        .withDisplayName("Lightning Probability")
-                                        .withLore(String.valueOf(lightningSettings.getLightning()))
-                                        .withNBTData(
-                                                c -> c.set(
-                                                        valueKey,
-                                                        PersistentDataType.INTEGER,
-                                                        lightningSettings.getLightning()))
-                                        .build(),
-                                12,
-                                ActionConsumer.getIntRange(valueKey, 0, 100),
-                                stack -> {
-                                    int fieldValue = DataContainerUtil.compute(stack, valueKey, PersistentDataType.INTEGER, s -> s);
-                                    lightningSettings.setLightning(fieldValue);
-                                })
-                );
-
-                actions.addAction(
-                        new ActionItem(
-                                ItemStackBuilder
-                                        .of(Material.BLAZE_POWDER)
-                                        .withDisplayName("Thunder Probability")
-                                        .withLore(String.valueOf(lightningSettings.getThunder()))
-                                        .withNBTData(
-                                                c -> c.set(
-                                                        valueKey,
-                                                        PersistentDataType.INTEGER,
-                                                        lightningSettings.getThunder()))
-                                        .build(),
-                                12,
-                                ActionConsumer.getIntRange(valueKey, 0, 100),
-                                stack -> {
-                                    int fieldValue = DataContainerUtil.compute(stack, valueKey, PersistentDataType.INTEGER, s -> s);
-                                    lightningSettings.setThunder(fieldValue);
-                                })
-                );
-                player.openInventory(inventory);
-                return true;
-            }
-
-            if ("shockwave".equalsIgnoreCase(field)) {
-                ShockwaveSettings shockwave = mobDeathActions.getShockwaveSettings();
-
-                Inventory inventory = Bukkit.createInventory(player, 56, "Manage Shockwave Settings");
-
-                InventoryActions actions = EldoUtilities.getInventoryActions().wrap(player, inventory, e -> {
-                    configuration.save();
-                    sendMobDeathActions(player, mobDeathActions);
-                });
-
-                // dont look at this O///O
-                NamespacedKey valueKey = new NamespacedKey(getPlugin(), "valueKey");
-                actions.addAction(
-                        ItemStackBuilder
-                                .of(Material.POTION)
-                                .withMetaValue(PotionMeta.class, m -> m.setColor(Color.RED))
-                                .withDisplayName("Effects")
-                                .build(),
-                        1,
-                        event -> {
-                            player.closeInventory();
-                            EldoUtilities.getDelayedActions().schedule(() -> {
-                                Inventory effectInventory = Bukkit.createInventory(player, 56, localizer().getMessage("Configure Shockwave Effects"));
-                                InventoryActions effectActions = EldoUtilities.getInventoryActions().wrap(player, effectInventory,
-                                        e -> {
-                                            configuration.save();
-                                            sendMobDeathActions(player, mobDeathActions);
-                                        });
-                                Map<PotionType, PotionEffectSettings> respawnEffects = shockwave.getShockwaveEffects();
-
-                                // this is always such a mess qwq
-                                int pos = 0;
-                                for (PotionType potionType : PotionType.values()) {
-                                    if (potionType.getEffectType() == null) continue;
-                                    pos++;
-                                    @Nullable PotionEffectSettings settings = respawnEffects.get(potionType);
-                                    effectActions.addAction(new ActionItem(
-                                            ItemStackBuilder
-                                                    .of(Material.POTION)
-                                                    .withDisplayName(potionType.getEffectType().getName())
-                                                    .withMetaValue(PotionMeta.class, m -> {
-                                                        m.setBasePotionData(new PotionData(PotionType.WATER_BREATHING));
-                                                        m.setColor(potionType.getEffectType().getColor());
-                                                    })
-                                                    .withNBTData(c ->
-                                                            c.set(valueKey, PersistentDataType.INTEGER, settings == null ? 0 : settings.getDuration()))
-                                                    .build(),
-                                            pos,
-                                            ActionConsumer.getIntRange(valueKey, 0, 600),
-                                            stack -> {
-                                                PersistentDataContainer container = stack.getItemMeta().getPersistentDataContainer();
-                                                int duration = container.get(valueKey, PersistentDataType.INTEGER);
-                                                if (duration == 0) return;
-                                                PotionMeta itemMeta = (PotionMeta) stack.getItemMeta();
-                                                PotionType type = itemMeta.getBasePotionData().getType();
-                                                shockwave.getShockwaveEffects().compute(type, (k, v) -> new PotionEffectSettings(type, duration));
-                                            }));
-                                }
-
-                            }, 1);
-                        },
-                        event -> {
-                        }
-                );
-
-                actions.addAction(
-                        ItemStackBuilder
-                                .of(Material.CLOCK)
-                                .withDisplayName("Min Effect Duration")
-                                .withNBTData(c -> c.set(valueKey, PersistentDataType.DOUBLE, shockwave.getMinDuration()))
-                                .build(),
-                        2,
-                        ActionConsumer.getDoubleRange(valueKey, 0, 60),
-                        item -> {
-                            Double aDouble = item.getItemMeta().getPersistentDataContainer().get(valueKey,
-                                    PersistentDataType.DOUBLE);
-                            shockwave.setMinDuration(aDouble);
-                        }
-                );
-
-                actions.addAction(
-                        ItemStackBuilder
-                                .of(Material.CLOCK)
-                                .withDisplayName("Range")
-                                .withNBTData(c -> c.set(valueKey, PersistentDataType.INTEGER, shockwave.getShockwaveRange()))
-                                .build(),
-                        3,
-                        ActionConsumer.getIntRange(valueKey, 0, 60),
-                        item -> {
-                            shockwave.setShockwaveRange(DataContainerUtil.getOrDefault(item, valueKey, PersistentDataType.INTEGER, 0));
-                        }
-                );
-
-                actions.addAction(
-                        ItemStackBuilder
-                                .of(Material.BLAZE_POWDER)
-                                .withDisplayName("POWER")
-                                .withNBTData(c -> c.set(valueKey, PersistentDataType.INTEGER, shockwave.getShockwavePower()))
-                                .build(),
-                        4,
-                        ActionConsumer.getIntRange(valueKey, 0, 60),
-                        item -> {
-                            shockwave.setShockwavePower(DataContainerUtil.getOrDefault(item, valueKey, PersistentDataType.INTEGER, 0));
-                        }
-                );
-
-                actions.addAction(
-                        ItemStackBuilder
-                                .of(Material.BLAZE_POWDER)
-                                .withDisplayName("Probability")
-                                .withNBTData(c -> c.set(valueKey, PersistentDataType.INTEGER, shockwave.getShockwaveProbability()))
-                                .build(),
-                        5,
-                        ActionConsumer.getIntRange(valueKey, 0, 100),
-                        item -> {
-                            shockwave.setShockwaveProbability(DataContainerUtil.getOrDefault(item, valueKey, PersistentDataType.INTEGER, 0));
-                        }
-                );
-
-
-                return true;
-            }
-        }
-
-        @Override
-        public @Nullable List<String> onTabComplete(@NotNull CommandSender sender, @NotNull Command command, @NotNull String alias, @NotNull String[] args) {
-            return super.onTabComplete(sender, command, alias, args);
-        }
-
-        private void sendMobDeathActions(Player player, MobDeathActions mobDeathActions) {
-
-        }
-    }
-
-    private static class ManagePlayerDeathActions extends EldoCommand {
-        private final Configuration configuration;
-        private final ConversationRequester conversationRequester;
-
-        public ManagePlayerDeathActions(Plugin plugin, Configuration configuration, ConversationRequester conversationRequester) {
-            super(plugin);
-            this.configuration = configuration;
-            this.conversationRequester = conversationRequester;
-        }
-
-        @Override
-        public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String label, @NotNull String[] args) {
-            Player player = (Player) sender;
-
-            World world = ArgumentUtils.getOrDefault(args, 0, ArgumentUtils::getWorld, player.getWorld());
-
-            if (world == null) {
-                messageSender().sendLocalizedError(sender, "error.invalidWorld");
-                return true;
-            }
-
-            PlayerDeathActions playerDeathActions = configuration.getWorldSettings(world).getDeathActionSettings().getPlayerDeathActions();
-
-            if (argumentsInvalid(sender, args, 2, "<$monster|player$> <$syntax.worldName$> [<$syntax.field$> <$syntax.value$>]")) {
-                return true;
-            }
-
-            String field = args[1];
-            String value = ArgumentUtils.getOrDefault(args, 2, "none");
-
-            if ("effects".equalsIgnoreCase(field)) {
-                // TODO
-                Inventory inventory = Bukkit.createInventory(player, 56, localizer().getMessage("Configure Respawn Effects"));
-                InventoryActions actions = EldoUtilities.getInventoryActions().wrap(player, inventory,
-                        e -> {
-                            configuration.save();
-                            sendPlayerDeathActions(playerDeathActions);
-                        });
-                Map<PotionType, PotionEffectSettings> respawnEffects = playerDeathActions.getRespawnEffects();
-
-                // this is always such a mess qwq
-                int pos = 0;
-                for (PotionType potionType : PotionType.values()) {
-                    if (potionType.getEffectType() == null) continue;
-                    pos++;
-                    @Nullable PotionEffectSettings settings = respawnEffects.get(potionType);
-                    NamespacedKey valueKey = new NamespacedKey(BloodNight.getInstance(), "value");
-                    actions.addAction(new ActionItem(
-                            ItemStackBuilder
-                                    .of(Material.POTION)
-                                    .withDisplayName(potionType.getEffectType().getName())
-                                    .withMetaValue(PotionMeta.class, m -> {
-                                        m.setBasePotionData(new PotionData(PotionType.WATER_BREATHING));
-                                        m.setColor(potionType.getEffectType().getColor());
-                                    })
-                                    .withNBTData(c ->
-                                            c.set(valueKey, PersistentDataType.INTEGER, settings == null ? 0 : settings.getDuration()))
-                                    .build(),
-                            pos,
-                            ActionConsumer.getIntRange(valueKey, 0, 600),
-                            stack -> {
-                                PersistentDataContainer container = stack.getItemMeta().getPersistentDataContainer();
-                                int duration = container.get(valueKey, PersistentDataType.INTEGER);
-                                if (duration == 0) return;
-                                PotionMeta itemMeta = (PotionMeta) stack.getItemMeta();
-                                PotionType type = itemMeta.getBasePotionData().getType();
-                                respawnEffects.compute(type, (k, v) -> new PotionEffectSettings(type, duration));
-                            }));
-                }
-                return true;
-            }
-
-            if ("commands".equalsIgnoreCase(field)) {
-                List<String> deathCommands = playerDeathActions.getDeathCommands();
-                Inventory inventory = Bukkit.createInventory(player, 56, "Manage Death Commands");
-                InventoryActions actions = EldoUtilities.getInventoryActions().wrap(player, inventory, e -> {
-                    configuration.save();
-                    sendPlayerDeathActions(playerDeathActions);
-                });
-
-                int pos = 0;
-                for (String deathCommand : deathCommands) {
-                    actions.addAction(
-                            new ActionItem(ItemStackBuilder
-                                    .of(Material.PAPER)
-                                    .withDisplayName(deathCommand)
-                                    .withLore("§2Left Click to change", "§cRight Click to remove.")
-                                    .build(),
-                                    pos,
-                                    e -> {
-                                        switch (e.getClick()) {
-                                            case LEFT:
-                                            case SHIFT_LEFT:
-                                                conversationRequester.requestInput(
-                                                        player,
-                                                        "Please enter command (Without /). Use {player} for player name.",
-                                                        s -> true, 0, i -> {
-                                                            deathCommands.set(pos, i);
-                                                            player.closeInventory();
-                                                        });
-                                                return;
-                                            case RIGHT:
-                                            case SHIFT_RIGHT:
-                                                deathCommands.remove(pos);
-                                                player.closeInventory();
-                                        }
-                                    },
-                                    e -> {
-                                    }));
-
-                }
-                return true;
-            }
-
-            if ("addCommand".equalsIgnoreCase(field)) {
-                if ("none".equalsIgnoreCase(field)) {
-                    messageSender().send(MessageChannel.CHAT, MessageType.ERROR, sender, "Please enter a command.");
-                    return true;
-                }
-                playerDeathActions.getDeathCommands().add(field);
-                configuration.save();
-                sendPlayerDeathActions(playerDeathActions);
-                return true;
-            }
-
-            if (TabCompleteUtil.isCommand(field, "loseExp", "loseInv")) {
-                OptionalInt optionalInt = Parser.parseInt(value);
-                if (!optionalInt.isPresent()) {
-                    messageSender().sendError(player, localizer().getMessage("error.invalidNumber"));
-                    return true;
-                }
-                if (invalidRange(sender, optionalInt.getAsInt(), 0, 100)) {
-                    return true;
-                }
-                if ("loseExp".equalsIgnoreCase(field)) {
-                    playerDeathActions.setLoseExpProbability(optionalInt.getAsInt());
-                }
-                if ("loseInventory".equalsIgnoreCase(field)) {
-                    playerDeathActions.setLoseInvProbability(optionalInt.getAsInt());
-                }
-                configuration.save();
-                return true;
-            }
-
-            if ("lightning".equalsIgnoreCase(field)) {
-                LightningSettings lightningSettings = playerDeathActions.getLightningSettings();
-
-                Inventory inventory = Bukkit.createInventory(player, 18, "Manage Lightning Settings");
-
-                InventoryActions actions = EldoUtilities.getInventoryActions().wrap(player, inventory, e -> {
-                    configuration.save();
-                    sendPlayerDeathActions(playerDeathActions);
-                });
-
-                NamespacedKey valueKey = new NamespacedKey(getPlugin(), "valueKey");
-                actions.addAction(
-                        new ActionItem(
-                                ItemStackBuilder
-                                        .of(Material.LEVER)
-                                        .withDisplayName("Lightning active")
-                                        .withLore(String.valueOf(lightningSettings.isDoLightning()))
-                                        .withNBTData(
-                                                c -> c.set(
-                                                        valueKey,
-                                                        PersistentDataType.BYTE,
-                                                        DataContainerUtil.booleanToByte(lightningSettings.isDoLightning())))
-                                        .build(),
-                                3,
-                                ActionConsumer.booleanToggle(valueKey),
-                                stack -> {
-                                    boolean fieldValue = DataContainerUtil.byteToBoolean(DataContainerUtil.compute(stack, valueKey, PersistentDataType.BYTE, s -> s));
-                                    lightningSettings.setDoLightning(fieldValue);
-                                })
-                );
-
-                actions.addAction(
-                        new ActionItem(
-                                ItemStackBuilder
-                                        .of(Material.LEVER)
-                                        .withDisplayName("Thunder active")
-                                        .withLore(String.valueOf(lightningSettings.isDoThunder()))
-                                        .withNBTData(
-                                                c -> c.set(
-                                                        valueKey,
-                                                        PersistentDataType.BYTE,
-                                                        DataContainerUtil.booleanToByte(lightningSettings.isDoThunder())))
-                                        .build(),
-                                5,
-                                ActionConsumer.booleanToggle(valueKey),
-                                stack -> {
-                                    boolean fieldValue = DataContainerUtil.byteToBoolean(DataContainerUtil.compute(stack, valueKey, PersistentDataType.BYTE, s -> s));
-                                    lightningSettings.setDoThunder(fieldValue);
-                                })
-                );
-
-                actions.addAction(
-                        new ActionItem(
-                                ItemStackBuilder
-                                        .of(Material.BLAZE_ROD)
-                                        .withDisplayName("Lightning Probability")
-                                        .withLore(String.valueOf(lightningSettings.getLightning()))
-                                        .withNBTData(
-                                                c -> c.set(
-                                                        valueKey,
-                                                        PersistentDataType.INTEGER,
-                                                        lightningSettings.getLightning()))
-                                        .build(),
-                                12,
-                                ActionConsumer.getIntRange(valueKey, 0, 100),
-                                stack -> {
-                                    int fieldValue = DataContainerUtil.compute(stack, valueKey, PersistentDataType.INTEGER, s -> s);
-                                    lightningSettings.setLightning(fieldValue);
-                                })
-                );
-
-                actions.addAction(
-                        new ActionItem(
-                                ItemStackBuilder
-                                        .of(Material.BLAZE_POWDER)
-                                        .withDisplayName("Thunder Probability")
-                                        .withLore(String.valueOf(lightningSettings.getThunder()))
-                                        .withNBTData(
-                                                c -> c.set(
-                                                        valueKey,
-                                                        PersistentDataType.INTEGER,
-                                                        lightningSettings.getThunder()))
-                                        .build(),
-                                12,
-                                ActionConsumer.getIntRange(valueKey, 0, 100),
-                                stack -> {
-                                    int fieldValue = DataContainerUtil.compute(stack, valueKey, PersistentDataType.INTEGER, s -> s);
-                                    lightningSettings.setThunder(fieldValue);
-                                })
-                );
-                player.openInventory(inventory);
-                return true;
-            }
-
-            if ("shockwave".equalsIgnoreCase(field)) {
-                ShockwaveSettings shockwave = playerDeathActions.getShockwaveSettings();
-
-                Inventory inventory = Bukkit.createInventory(player, 56, "Manage Shockwave Settings");
-
-                InventoryActions actions = EldoUtilities.getInventoryActions().wrap(player, inventory, e -> {
-                    configuration.save();
-                    sendPlayerDeathActions(playerDeathActions);
-                });
-
-                // dont look at this O///O
-                NamespacedKey valueKey = new NamespacedKey(getPlugin(), "valueKey");
-                actions.addAction(
-                        ItemStackBuilder
-                                .of(Material.POTION)
-                                .withMetaValue(PotionMeta.class, m -> m.setColor(Color.RED))
-                                .withDisplayName("Effects")
-                                .build(),
-                        1,
-                        event -> {
-                            player.closeInventory();
-                            EldoUtilities.getDelayedActions().schedule(() -> {
-                                Inventory effectInventory = Bukkit.createInventory(player, 56, localizer().getMessage("Configure Shockwave Effects"));
-                                InventoryActions effectActions = EldoUtilities.getInventoryActions().wrap(player, effectInventory,
-                                        e -> {
-                                            configuration.save();
-                                            sendPlayerDeathActions(playerDeathActions);
-                                        });
-                                Map<PotionType, PotionEffectSettings> respawnEffects = shockwave.getShockwaveEffects();
-
-                                // this is always such a mess qwq
-                                int pos = 0;
-                                for (PotionType potionType : PotionType.values()) {
-                                    if (potionType.getEffectType() == null) continue;
-                                    pos++;
-                                    @Nullable PotionEffectSettings settings = respawnEffects.get(potionType);
-                                    effectActions.addAction(new ActionItem(
-                                            ItemStackBuilder
-                                                    .of(Material.POTION)
-                                                    .withDisplayName(potionType.getEffectType().getName())
-                                                    .withMetaValue(PotionMeta.class, m -> {
-                                                        m.setBasePotionData(new PotionData(PotionType.WATER_BREATHING));
-                                                        m.setColor(potionType.getEffectType().getColor());
-                                                    })
-                                                    .withNBTData(c ->
-                                                            c.set(valueKey, PersistentDataType.INTEGER, settings == null ? 0 : settings.getDuration()))
-                                                    .build(),
-                                            pos,
-                                            ActionConsumer.getIntRange(valueKey, 0, 600),
-                                            stack -> {
-                                                PersistentDataContainer container = stack.getItemMeta().getPersistentDataContainer();
-                                                int duration = container.get(valueKey, PersistentDataType.INTEGER);
-                                                if (duration == 0) return;
-                                                PotionMeta itemMeta = (PotionMeta) stack.getItemMeta();
-                                                PotionType type = itemMeta.getBasePotionData().getType();
-                                                shockwave.getShockwaveEffects().compute(type, (k, v) -> new PotionEffectSettings(type, duration));
-                                            }));
-                                }
-
-                            }, 1);
-                        },
-                        event -> {
-                        }
-                );
-
-                actions.addAction(
-                        ItemStackBuilder
-                                .of(Material.CLOCK)
-                                .withDisplayName("Min Effect Duration")
-                                .withNBTData(c -> c.set(valueKey, PersistentDataType.DOUBLE, shockwave.getMinDuration()))
-                                .build(),
-                        2,
-                        ActionConsumer.getDoubleRange(valueKey, 0, 60),
-                        item -> {
-                            Double aDouble = item.getItemMeta().getPersistentDataContainer().get(valueKey,
-                                    PersistentDataType.DOUBLE);
-                            shockwave.setMinDuration(aDouble);
-                        }
-                );
-
-                actions.addAction(
-                        ItemStackBuilder
-                                .of(Material.CLOCK)
-                                .withDisplayName("Range")
-                                .withNBTData(c -> c.set(valueKey, PersistentDataType.INTEGER, shockwave.getShockwaveRange()))
-                                .build(),
-                        3,
-                        ActionConsumer.getIntRange(valueKey, 0, 60),
-                        item -> {
-                            shockwave.setShockwaveRange(DataContainerUtil.getOrDefault(item, valueKey, PersistentDataType.INTEGER, 0));
-                        }
-                );
-
-                actions.addAction(
-                        ItemStackBuilder
-                                .of(Material.BLAZE_POWDER)
-                                .withDisplayName("POWER")
-                                .withNBTData(c -> c.set(valueKey, PersistentDataType.INTEGER, shockwave.getShockwavePower()))
-                                .build(),
-                        4,
-                        ActionConsumer.getIntRange(valueKey, 0, 60),
-                        item -> {
-                            shockwave.setShockwavePower(DataContainerUtil.getOrDefault(item, valueKey, PersistentDataType.INTEGER, 0));
-                        }
-                );
-
-                actions.addAction(
-                        ItemStackBuilder
-                                .of(Material.BLAZE_POWDER)
-                                .withDisplayName("Probability")
-                                .withNBTData(c -> c.set(valueKey, PersistentDataType.INTEGER, shockwave.getShockwaveProbability()))
-                                .build(),
-                        5,
-                        ActionConsumer.getIntRange(valueKey, 0, 100),
-                        item -> {
-                            shockwave.setShockwaveProbability(DataContainerUtil.getOrDefault(item, valueKey, PersistentDataType.INTEGER, 0));
-                        }
-                );
-                return true;
-            }
-
-            sendPlayerDeathActions(playerDeathActions);
-            return true;
-        }
-
-        private void sendPlayerDeathActions(PlayerDeathActions playerDeathActions) {
-
-        }
-
-
-        @Override
-        public @Nullable List<String> onTabComplete(@NotNull CommandSender sender, @NotNull Command command, @NotNull String alias, @NotNull String[] args) {
-            return super.onTabComplete(sender, command, alias, args);
-        }
     }
 }
