+++ BloodNight-core/src/main/resources/messages.properties
@@ -124,3 +124,29 @@ syntax.value=value
 syntax.worldName=worldName
 value.second=second
 value.seconds=seconds
+manageDeathActions.player.title=Manage Player Death Settings
+manageDeathActions.monster.title=Manage Monster Death Settings
+field.lightningSettings=Lightning Settings
+field.shockwaveSettings=Shockwave Settings
+field.deathCommands=Death Commands (%COUNT%)
+field.respawnEffect=Respawn Effects
+field.loseInventory=Lose Inventory
+field.loseExperience=Lose Experience
+action.add=add
+manageDeathActions.inventory.lightning.title=Manage Lightning Settings
+manageDeathActions.inventory.lightning.lightningActive=Lightning Active
+manageDeathActions.inventory.lightning.lightningProb=Lightning Probability
+manageDeathActions.inventory.lightning.thunderActive=Thunder Active
+manageDeathActions.inventory.lightning.thunderProb=Thunder Probability
+manageDeathActions.inventory.shockwave.title=Manage Shockwave Settings
+manageDeathActions.inventory.shockwave.effects=Manage Shockwave Effects
+manageDeathActions.inventory.shockwave.minEffectDuration=Min Effect Duration
+manageDeathActions.inventory.respawnEffects.title=Manage Respawn Effects
+manageDeathActions.phrase.commandAdded=Command %CMD% added.
+field.range=Range
+field.power=Power
+phrase.leftClickChange=Left Click to change
+phrase.rightClickRemove=Right Click to remove
+phrase.commandPlayer=Please enter command (Without /). Use {player} for player name.
+syntax.commandPlayer=Please enter a command. Use {player} as playername placeholder
+error.noCommand=Please enter a command.
